♻️: seal the mutable (Set) Anima API
To prevent misuse breaking Bevy change detection mechanisms.
