+++ game_client/crates/anima/src/lib.rs
@@ -6,8 +6,10 @@ mod enable_anima;
 mod meta;
 mod plugins;
 mod ted;
+mod ted_seal;
 mod transform;
 mod trs;
+mod trs_seal;
 
 pub use anima::*;
 use animate_anima::*;
@@ -17,5 +19,7 @@ pub use enable_anima::*;
 use meta::*;
 pub use plugins::*;
 pub use ted::*;
+use ted_seal::*;
 use transform::*;
 pub use trs::*;
+use trs_seal::*;
